Refactor import section in unit tests.
