+++ tests/storage/aws_test.py
@@ -24,9 +24,7 @@ import boto3
 import pytest
 from typing_extensions import Self
 
-from esrally.storage._adapter import Head
-from esrally.storage._config import StorageConfig
-from esrally.storage._range import rangeset
+from esrally.storage import Head, StorageConfig, rangeset
 from esrally.storage.aws import S3Adapter, S3Client, head_from_response
 from esrally.utils.cases import cases
